+++ App/Drivers/Payments/KlarnaPaymentDriver.php
@@ -332,7 +332,7 @@ class KlarnaPaymentDriver extends PaymentDriver
 
     private function apiUrl(string $path): string
     {
-        return rtrim((string) $this->setting('API_BASE'), '/') . $path;
+        return rtrim((string) $this->setting('API_BASE', $this->setting('BASE_URL', '')), '/') . $path;
     }
 
     /**